[Artwork] Wrap tabs in Theme
