--- src/desktop/components/react/stitch_components/ReactionTooltipQuestion.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import { Help } from 'reaction/Assets/Icons/Help'
-import { Tooltip } from 'reaction/Components/Tooltip'
-
-export const ReactionTooltipQuestion = props => {
-  return (
-    <Tooltip {...props}>
-      <Help />
-    </Tooltip>
-  )
-}
